+++ cphasing/mapper.py
@@ -307,7 +307,8 @@ class ChromapMapper:
 
 class PoreCMapper:
     def __init__(self, reference, read, pattern="GATC", 
-                    k=15, w=10, min_quality=1, realign=False,
+                    k=15, w=10, min_quality=1, 
+                    min_identity=0.75, min_length=30, realign=False,
                     additional_arguments=("-x", "map-ont"), outprefix=None,
                     threads=4, path='minimap2', log_dir='logs',
                     force=False):
@@ -347,6 +348,8 @@ class PoreCMapper:
 
         self.threads = threads
         self.min_quality = min_quality
+        self.min_identity = min_identity
+        self.min_length = min_length
         self.k = k 
         self.w = w
         self.force = force
@@ -446,7 +449,9 @@ class PoreCMapper:
         if self.pattern:
             cmd = ["cphasing-rs", "paf2porec", f"{paf}", "-b",
                    f"{self.reference.stem}.slope.{self.pattern}.bed", "-q", 
-                    f"{self.min_quality}", "-o", f"{self.outporec}",
+                    f"{self.min_quality}", "-l", f"{self.min_length}", 
+                    "-p", f"{self.min_identity}",
+                    "-o", f"{self.outporec}",
                 ]
         else:
             cmd = ["cphasing-rs", "paf2porec", f"{paf}", "-q",